Set up an experiment where a Dyna-Q agent faces a changing environment
This is heavily based on the dyna-Q example elsewhere in this repository.

In this version, the maze environment can change after some fixed number of
episodes, and the agent needs to respond accordingly.

As it's currently set up, the agent does nothing in particular to handle the
change, and we see an expected spike in steps per episode where the agent
suddenly gets completely lost. By virtue of the epsilon-greedy action
selection, it does eventually find the new way around the maze and continue.

Follow-up changes will experiment with two possible solutions to this problem.
This is exercise 8.4 in Sutton & Barto RL: An Introduction.