Implement trajectory-based sampling
This actually seems to perform worse in the simple maze environment.

I made the maze slightly bigger, but the results were the same. It seems that
it's not as useful when the state space is this small, and making the maze
larger still just made it take an absurdly long time to train.

However, the implementation works!