Add random rewards and implement basic expected updates
The 'expected updates' are really only averages of the random rewards received
to avoid having to significantly change the dynamics of the environment.

The new expected update method keeps track of expected rewards in the model and
plans accordingly.

The random rewards were enough to break the existing planning methods, but the
new one with expected updates succeeds.