Fix the algorithm and demonstrate performance
Previously the softmax was using x instead of e^x, making the sum of all
preferences zero and forcing us to resort to giving equal weight to all
actions. This meant performance was purely random.

This change fixes that and refactors the code, also adding performance graphing
to demonstrate that the algorithm quickly learns a good policy.