Fix bugs with the policy playground
It turns out that Python imports are bizarre:
https://stackoverflow.com/questions/2282369/how-do-i-get-the-module-name-of-an-objects-class-definition-rather-than-the-mod

Actions created by the track environment weren't matching Actions created by
the policy. Making one of the imports of the definition of Action an explicit
relative import solved the problem.

It's now possible to train up an agent and then set it running in the
playground:

python sarsa_lambda.py
python policy_playground.py